ref(texture_api): update to the texture api Replaces deprecated GT_Textures objects with the GT TextureBuilder API.

+++ src/main/java/gtPlusPlus/core/entity/monster/EntityGiantChickenBase.java
@@ -255,13 +255,13 @@ public class EntityGiantChickenBase extends EntityChicken {
         this.motionY += 0.068999998688697815D;
         if (this.isPotionActive(Potion.jump)) {
             this.motionY +=
-                    (double) ((float) (this.getActivePotionEffect(Potion.jump).getAmplifier() + 1) * 0.5F);
+                    (float) (this.getActivePotionEffect(Potion.jump).getAmplifier() + 1) * 0.5F;
         }
 
         if (this.isSprinting()) {
             float f = this.rotationYaw * 0.017453292F;
-            this.motionX -= (double) (MathHelper.sin(f) * 0.2F);
-            this.motionZ += (double) (MathHelper.cos(f) * 0.2F);
+            this.motionX -= MathHelper.sin(f) * 0.2F;
+            this.motionZ += MathHelper.cos(f) * 0.2F;
         }
 
         this.isAirBorne = true;
